+++ CHANGELOG.md
@@ -17,6 +17,7 @@
 
 - `lib`: Remove variable PROGRAM from lib part
 - `lib`: Remove logging functions
+- `lib/tman`: Delete library functions' return code
 
 ### Deprecated